+++ src/lib/src/client/index.ts
@@ -1,7 +1,15 @@
 export { default as TableHandler } from './TableHandler.svelte'
 export { check } from './Comparator'
 import type { UUID } from 'crypto'
-import type { Internationalization, Field } from '$lib/src/shared'
+import type { Internationalization, Field }     from '$lib/src/shared'
+export type { default as SearchBuilder }        from './builders/SearchBuilder.svelte'
+export type { default as CalculationBuilder }   from './builders/CalculationBuilder.svelte'
+export type { default as CSVBuilder }           from './builders/CSVBuilder.svelte'
+export type { default as FilterBuilder }        from './builders/FilterBuilder.svelte'
+export type { default as RecordFilterBuilder }  from './builders/RecordFilterBuilder.svelte'
+export type { default as SortBuilder }          from './builders/SortBuilder.svelte'
+export { default as RecordFilter }              from './builders/RecordFilterBuilder.svelte'
+
 export {
     Datatable,
     Search,
@@ -15,12 +23,12 @@ export {
     type Field,
     type Internationalization,
     type ColumnView,
-    type TableHandlerInterface
+    type TableHandlerInterface,
 } from '$lib/src/shared'
 
 
 export type Filter<Row> = {
-    callback: (row: Row) => unknown
+    callback: (row: $state.Snapshot<Row>) => unknown
     id: UUID,
     value?: unknown
     check?: Check